Move schema to be inline
